fixbug: ds will coredump when decrease quota. see http://rb.corp.taobao.com/r/34438/
git-svn-id: http://code.taobao.org/svn/tair/trunk@942 ed5bfd34-b61d-4ad5-b956-f82ce90ca771

+++ src/storage/mdb/mem_cache.hpp
@@ -139,7 +139,7 @@ namespace tair {
     bool is_quota_exceed(int area);
     void calc_slab_balance_info(std::map<int, int > &adjust_info);
     void balance_slab_done();
-    void keep_area_quota(int area, int exceed);
+    void keep_area_quota(int area, uint64_t exceed);
 
     struct mdb_cache_info
     {
